+++ pkg/backup/delete.go
@@ -11,7 +11,7 @@ import (
 	"time"
 
 	"github.com/AlexAkulov/clickhouse-backup/pkg/clickhouse"
-	"github.com/AlexAkulov/clickhouse-backup/pkg/new_storage"
+	"github.com/AlexAkulov/clickhouse-backup/pkg/storage"
 
 	apexLog "github.com/apex/log"
 )
@@ -134,7 +134,7 @@ func RemoveBackupRemote(cfg *config.Config, backupName string) error {
 	}
 	backupName = CleanBackupNameRE.ReplaceAllString(backupName, "")
 
-	bd, err := new_storage.NewBackupDestination(cfg, false)
+	bd, err := storage.NewBackupDestination(cfg, false)
 	if err != nil {
 		return err
 	}