+++ pkg/custom/list_custom.go
@@ -4,14 +4,14 @@ import (
 	"encoding/json"
 	"fmt"
 	"github.com/AlexAkulov/clickhouse-backup/pkg/config"
-	"github.com/AlexAkulov/clickhouse-backup/pkg/new_storage"
+	"github.com/AlexAkulov/clickhouse-backup/pkg/storage"
 	"github.com/AlexAkulov/clickhouse-backup/pkg/utils"
 	"github.com/apex/log"
 	"strings"
 	"time"
 )
 
-func List(cfg *config.Config) ([]new_storage.Backup, error) {
+func List(cfg *config.Config) ([]storage.Backup, error) {
 	if cfg.Custom.ListCommand == "" {
 		return nil, fmt.Errorf("CUSTOM_LIST_COMMAND is not defined")
 	}
@@ -23,7 +23,7 @@ func List(cfg *config.Config) ([]new_storage.Backup, error) {
 	out, err := utils.ExecCmdOut(cfg.Custom.CommandTimeoutDuration, args[0], args[1:]...)
 	if err == nil {
 		outLines := strings.Split(strings.TrimRight(out, "\n"), "\n")
-		backupList := make([]new_storage.Backup, len(outLines))
+		backupList := make([]storage.Backup, len(outLines))
 		for i, line := range outLines {
 			if len(line) > 0 {
 				if err = json.Unmarshal([]byte(line), &backupList[i]); err != nil {